% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covs.R
\docType{data}
\name{covs}
\alias{covs}
\title{Sample covairate data to demonstrate sample designs}
\format{
An object of class `"DataFrame"`.
}
\source{
https://doi.org/10.1016/j.geoderma.2022.116054
}
\usage{
data(covs)
}
\description{
Covariate data for the Smeltzer farm near Guelph, Ontario
}
\keyword{datasets}
