% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm.covs.R
\name{dsm.covs}
\alias{dsm.covs}
\title{Creating_Continuous_Covariates for digital soil mapping}
\usage{
dsm.covs(x, wdir = NULL, cpus = NULL)
}
\arguments{
\item{x}{SpatRaster}

\item{wdir}{character User can specify the working directory, defaults to tempdir()}

\item{cpus, }{numeric Used to reduce the number of cores used by RSAGA}
}
\value{
SpatRaster Returns a SPatRaster with 55 covariates pointing to the rasters in working directory
}
\description{
This function creates a series of 55 environmental
covariates derived from a digital elevation model.
These are commonly used covariates in digital soil mapping.
These are created using SAGA-GIS and WhiteboxTools through
R implementations in RSAGA and whitebox packages
}
\examples{
#let us use the Keene digital terrain model as an example
data(keene)
keene<- terra::rast(keene, type="xyz")
terra::crs(keene)<- "epsg:26917"
covs<- dsm.covs(keene)
}
