% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_sample.R
\name{opt_sample}
\alias{opt_sample}
\title{Determine optimal sample size for the conditioned Latin hypercube
or feature (covariate) coverage sampling algorithms
using three different divergence metrics:
Kullback-Leibler Divergence, Jenson-Shannon Divergence,
and Jenson-Shannon Distance}
\usage{
opt_sample(
  alg = "clhs",
  s_min,
  s_max,
  s_step = 20,
  s_reps = 2,
  bins = 30,
  covs,
  clhs_iter = 10000,
  cpus = NULL,
  conf = 0.95
)
}
\arguments{
\item{alg, }{character, either "clhs" or "fcs", default is "clhs"}

\item{s_min}{numeric, minimum sample size to be tested}

\item{s_max}{numeric, maximum sample size to be tested}

\item{s_step}{numeric, sample size step for testing, default is 20}

\item{s_reps}{numeric, number of repetitions of the sample design, default is 2, must be >1}

\item{bins}{numeric, number of bins for the divergence metric calculation, default is 30}

\item{covs}{data frame, environmental covariates to be used for the sample size determination}

\item{clhs_iter}{numeric, number of iterations to use for the clhs, default is 10000}

\item{cpus}{numeric, number of CPUs to be used for parallel processing, default is 1/2 of virtual CPUs}

\item{conf}{numeric, confidence used to select optimal sample size from the CDF, default is 0.95}
}
\value{
list with 3 dataframes, 1 list, and 2 figures are plotted
Dataframe 1: Dataframe with optimal sample size based on the three divergence metrics
Dataframe 2: List of vectors showing the index of rows for each sample plan evaluated
Dataframe 3: Dataframe providing mean and standard deviation of divergence metrics across the repetitions
Dataframe 4: Dataframe showing all raw data - divergence metrics for each covariate at each sample size
Figure 1: plots of exponential decay of mean divergence metrics across all covariates with increasing sample size
Figure 2: plots of exponential decay of normalized divergence metrics for individual covariates with increasing sample size
}
\description{
This script is based on the publication:
Divergence metrics for determining optimal training sample size in digital soil mapping
DOI: https://doi.org/10.1016/j.geoderma.2023.116553
}
\examples{
# import the example data and run with 4 reps
data(covs)
out<- opt_sample(alg="clhs", s_min=10, s_max=400, s_step=20, s_reps=4, covs= covs, clhs_iter=100, cpus=NULL, conf=0.95)

}
